+++ src/PhpWord/Writer/ODText/Element/Text.php
@@ -9,10 +9,6 @@
 
 namespace PhpOffice\PhpWord\Writer\ODText\Element;
 
-use PhpOffice\PhpWord\Shared\String;
-use PhpOffice\PhpWord\Style\Font;
-use PhpOffice\PhpWord\Style\Paragraph;
-
 /**
  * Text element writer
  *